Split guild types
