+++ packages/types/src/discord/gateway.ts
@@ -6,14 +6,10 @@
 
 import type {
   DiscordAvatarDecorationData,
-  DiscordIntegration,
-  DiscordMember,
-  DiscordMemberWithUser,
   DiscordReactionType,
   DiscordRole,
   DiscordSoundboardSound,
   DiscordSticker,
-  DiscordUnavailableGuild,
   DiscordUser,
 } from '../discord.js'
 import type { GatewayEventNames, TargetTypes } from '../shared.js'
@@ -21,6 +17,7 @@ import type { DiscordApplication } from './applications.js'
 import type { AutoModerationTriggerTypes, DiscordAutoModerationAction } from './autoModeration.js'
 import type { DiscordChannel, DiscordThreadMember } from './channels.js'
 import type { DiscordEmoji } from './emojis.js'
+import type { DiscordIntegration, DiscordMember, DiscordMemberWithUser, DiscordUnavailableGuild } from './guilds.js'
 
 /** https://discord.com/developers/docs/events/gateway#list-of-intents */
 export enum GatewayIntents {
