+++ packages/types/src/index.ts
@@ -6,6 +6,7 @@ export * from './discord/channels.js'
 export * from './discord/emojis.js'
 export * from './discord/entitlements.js'
 export * from './discord/gateway.js'
+export * from './discord/guilds.js'
 export * from './discord/interactions.js'
 export * from './discord/reference.js'
 export * from './discord/webhookEvents.js'
