+++ packages/types/src/shared.ts
@@ -29,67 +29,6 @@ export enum UserFlags {
   ActiveDeveloper = 1 << 22,
 }
 
-export enum MemberFlags {
-  /**
-   * Member has left and rejoined the guild
-   *
-   * @remarks
-   * This value is not editable
-   */
-  DidRejoin = 1 << 0,
-  /**
-   * Member has completed onboarding
-   *
-   * @remarks
-   * This value is not editable
-   */
-  CompletedOnboarding = 1 << 1,
-  /** Member is exempt from guild verification requirements */
-  BypassesVerification = 1 << 2,
-  /**
-   * Member has started onboarding
-   *
-   * @remarks
-   * This value is not editable
-   */
-  StartedOnboarding = 1 << 3,
-  /**
-   * Member is a guest and can only access the voice channel they were invited to
-   *
-   * @remarks
-   * This value is not editable
-   */
-  IsGuest = 1 << 4,
-  /**
-   * Member has started Server Guide new member actions
-   *
-   * @remarks
-   * This value is not editable
-   */
-  StartedHomeActions = 1 << 5,
-  /**
-   * Member has completed Server Guide new member actions
-   *
-   * @remarks
-   * This value is not editable
-   */
-  CompletedHomeActions = 1 << 6,
-  /**
-   * Member's username, display name, or nickname is blocked by AutoMod
-   *
-   * @remarks
-   * This value is not editable
-   */
-  AutomodQuarantinedUsername = 1 << 7,
-  /**
-   * Member has dismissed the DM settings upsell
-   *
-   * @remarks
-   * This value is not editable
-   */
-  DmSettingsUpsellAcknowledged = 1 << 9,
-}
-
 /** https://discord.com/developers/docs/topics/permissions#role-object-role-flags */
 export enum RoleFlags {
   None,
@@ -139,12 +78,6 @@ export enum MessageFlags {
   IsVoiceMessage = 1 << 13,
 }
 
-/** https://discord.com/developers/docs/resources/guild#integration-object-integration-expire-behaviors */
-export enum IntegrationExpireBehaviors {
-  RemoveRole,
-  Kick,
-}
-
 /** https://discord.com/developers/docs/topics/teams#data-models-membership-state-enum */
 export enum TeamMembershipStates {
   Invited = 1,
@@ -173,140 +106,6 @@ export enum WebhookTypes {
 
 /** https://discord.com/developers/docs/resources/channel#embed-object-embed-types */
 export type EmbedTypes = 'rich' | 'image' | 'video' | 'gifv' | 'article' | 'link' | 'poll_result'
-
-/** https://discord.com/developers/docs/resources/guild#guild-object-default-message-notification-level */
-export enum DefaultMessageNotificationLevels {
-  /** Members will receive notifications for all messages by default */
-  AllMessages,
-  /** Members will receive notifications only for messages that \@mention them by default */
-  OnlyMentions,
-}
-
-/** https://discord.com/developers/docs/resources/guild#guild-object-explicit-content-filter-level */
-export enum ExplicitContentFilterLevels {
-  /** Media content will not be scanned */
-  Disabled,
-  /** Media content sent by members without roles will be scanned */
-  MembersWithoutRoles,
-  /** Media content sent by all members will be scanned */
-  AllMembers,
-}
-
-/** https://discord.com/developers/docs/resources/guild#guild-object-verification-level */
-export enum VerificationLevels {
-  /** Unrestricted */
-  None,
-  /** Must have verified email on account */
-  Low,
-  /** Must be registered on Discord for longer than 5 minutes */
-  Medium,
-  /** Must be a member of the server for longer than 10 minutes */
-  High,
-  /** Must have a verified phone number */
-  VeryHigh,
-}
-
-/** https://discord.com/developers/docs/resources/guild#guild-object-guild-features */
-export enum GuildFeatures {
-  /** Guild has access to set an invite splash background */
-  InviteSplash = 'INVITE_SPLASH',
-  /** Guild has access to set 384 kbps bitrate in voice (previously VIP voice servers) */
-  VipRegions = 'VIP_REGIONS',
-  /** Guild has access to set a vanity URL */
-  VanityUrl = 'VANITY_URL',
-  /** Guild is verified */
-  Verified = 'VERIFIED',
-  /** Guild is partnered */
-  Partnered = 'PARTNERED',
-  /** Guild can enable welcome screen, Membership Screening, stage channels and discovery, and receives community updates */
-  Community = 'COMMUNITY',
-  /** Guild has enabled monetization. */
-  CreatorMonetizableProvisional = 'CREATOR_MONETIZABLE_PROVISIONAL',
-  /** Guild has enabled the role subscription promo page. */
-  CreatorStorePage = 'CREATOR_STORE_PAGE',
-  /** Guild has been set as a support server on the App Directory */
-  DeveloperSupportServer = 'DEVELOPER_SUPPORT_SERVER',
-  /** Guild has access to create news channels */
-  News = 'NEWS',
-  /** Guild is able to be discovered in the directory */
-  Discoverable = 'DISCOVERABLE',
-  /** Guild is able to be featured in the directory */
-  Featurable = 'FEATURABLE',
-  /** Guild has access to set an animated guild icon */
-  AnimatedIcon = 'ANIMATED_ICON',
-  /** Guild has access to set a guild banner image */
-  Banner = 'BANNER',
-  /** Guild has enabled the welcome screen */
-  WelcomeScreenEnabled = 'WELCOME_SCREEN_ENABLED',
-  /** Guild has enabled [Membership Screening](https://discord.com/developers/docs/resources/guild#membership-screening-object) */
-  MemberVerificationGateEnabled = 'MEMBER_VERIFICATION_GATE_ENABLED',
-  /** Guild has increased custom soundboard sound slots. */
-  MoreSoundboard = 'MORE_SOUNDBOARD',
-  /** Guild can be previewed before joining via Membership Screening or the directory */
-  PreviewEnabled = 'PREVIEW_ENABLED',
-  /** Guild has enabled ticketed events */
-  TicketedEventsEnabled = 'TICKETED_EVENTS_ENABLED',
-  /** Guild has increased custom sticker slots */
-  MoreStickers = 'MORE_STICKERS',
-  /** Guild is able to set role icons */
-  RoleIcons = 'ROLE_ICONS',
-  /** Guild has role subscriptions that can be purchased. */
-  RoleSubscriptionsAvailableForPurchase = 'ROLE_SUBSCRIPTIONS_AVAILABLE_FOR_PURCHASE',
-  /** Guild has enabled role subscriptions. */
-  RoleSubscriptionsEnabled = 'ROLE_SUBSCRIPTIONS_ENABLED',
-  /** Guild has created soundboard sounds. */
-  Soundboard = 'SOUNDBOARD',
-  /** Guild has set up auto moderation rules */
-  AutoModeration = 'AUTO_MODERATION',
-  /** Guild has paused invites, preventing new users from joining */
-  InvitesDisabled = 'INVITES_DISABLED',
-  /** Guild has access to set an animated guild banner image */
-  AnimatedBanner = 'ANIMATED_BANNER',
-  /** Guild has disabled alerts for join raids in the configured safety alerts channel */
-  RaidAlertsDisabled = 'RAID_ALERTS_DISABLED',
-  /** Guild is using the old permissions configuration behavior */
-  ApplicationCommandPermissionsV2 = 'APPLICATION_COMMAND_PERMISSIONS_V2',
-}
-
-/** https://discord.com/developers/docs/resources/guild#guild-object-mfa-level */
-export enum MfaLevels {
-  /** Guild has no MFA/2FA requirement for moderation actions */
-  None,
-  /** Guild has a 2FA requirement for moderation actions */
-  Elevated,
-}
-
-/** https://discord.com/developers/docs/resources/guild#guild-object-system-channel-flags */
-export enum SystemChannelFlags {
-  /** Suppress member join notifications */
-  SuppressJoinNotifications = 1 << 0,
-  /** Suppress server boost notifications */
-  SuppressPremiumSubscriptions = 1 << 1,
-  /** Suppress server setup tips */
-  SuppressGuildReminderNotifications = 1 << 2,
-  /** Hide member join sticker reply buttons */
-  SuppressJoinNotificationReplies = 1 << 3,
-}
-
-/** https://discord.com/developers/docs/resources/guild#guild-object-premium-tier */
-export enum PremiumTiers {
-  /** Guild has not unlocked any Server Boost perks */
-  None,
-  /** Guild has unlocked Server Boost level 1 perks */
-  Tier1,
-  /** Guild has unlocked Server Boost level 2 perks */
-  Tier2,
-  /** Guild has unlocked Server Boost level 3 perks */
-  Tier3,
-}
-
-/** https://discord.com/developers/docs/resources/guild#guild-object-guild-nsfw-level */
-export enum GuildNsfwLevel {
-  Default,
-  Explicit,
-  Safe,
-  AgeRestricted,
-}
 
 /** https://discord.com/developers/docs/resources/channel#message-object-message-types */
 export enum MessageTypes {
